requirements.txt
Pillow
numpy
scikit-image
scikit-learn 